fix: reuse Swap for Market token trading (OK-61581)
