+++ packages/kit/src/views/Market/MarketDetailV2/MarketDetailV2.tsx
@@ -97,7 +97,7 @@ function MarketDetail({
     | ITabMarketParamList[ETabMarketRoutes.MarketStockDetail]
     | ITabMarketParamList[ETabMarketRoutes.MarketNativeDetail];
 
-  const { selectedTokenVariant } = useStockDetail();
+  const { isStockRoute, selectedTokenVariant } = useStockDetail();
   const network =
     selectedTokenVariant?.networkId ??
     ('network' in params ? params.network : '') ??
@@ -179,8 +179,9 @@ function MarketDetail({
     preloadMarketDetailV2BodyModules({
       layout: isDesktopLayout ? 'desktop' : 'mobile',
       includeHeavyModules: true,
+      isStockRoute,
     });
-  }, [isDesktopLayout]);
+  }, [isDesktopLayout, isStockRoute]);
 
   return (
     <BtcMetadataProvider>
