+++ packages/kit/src/views/Market/MarketHomeV2/components/MarketStockList/hooks/useToMarketStockDetailPage.ts
@@ -33,6 +33,7 @@ export function useToMarketStockDetailPage() {
       const preloadPromise = preloadMarketDetailV2Page({
         includeBodyModules: true,
         includeHeavyModules: true,
+        isStockRoute: true,
         layout: preloadLayout,
       });
       tokenDetailActions.current.clearTokenDetail();
